modified Main.tsx; moved email validation to separate file
