+++ web-client/src/pages/session/login.tsx
@@ -1,19 +1,14 @@
 import React from 'react';
-
 import axios from 'axios';
+
 import { useRouter } from 'next/router';
-import { useAsyncEffect } from 'use-async-effect';
+
+import { UserConstants } from './types';
 
 import { Meta } from '../../layout/Meta';
 import { Main } from '../../templates/Main';
 import { AppConfig } from '../../utils/AppConfig';
-import { UserConstants } from './types';
-
-function validEmail(email: string) {
-  const regex =
-    /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-  return email.match(regex);
-}
+import { validEmail } from '../../utils/Email';
 
 const LoginPage = () => {
   const [email, setEmail] = React.useState('');
@@ -23,31 +18,6 @@ const LoginPage = () => {
 
   function redirectToHomepage() {
     router.push('/');
-  }
-
-  async function checkLoggedIn() {
-    const userJson = localStorage.getItem(UserConstants.CURRENT);
-    const jwt = localStorage.getItem(UserConstants.JWT);
-    if (userJson == null || jwt == null) {
-      return false;
-    }
-
-    const user = JSON.parse(userJson);
-    try {
-      await axios.get(`http://localhost:4000/v1/users/${user.id}`, {
-        headers: {
-          Authorization: jwt,
-        },
-      });
-      alert('You are already logged in!');
-    } catch (err) {
-      alert('Your JWT token has expired! Logging out...');
-      localStorage.removeItem(UserConstants.CURRENT);
-      localStorage.removeItem(UserConstants.JWT);
-    } finally {
-      redirectToHomepage();
-    }
-    return true;
   }
 
   async function onLoginClick() {
@@ -75,8 +45,6 @@ const LoginPage = () => {
       alert('Invalid email or password!');
     }
   }
-
-  useAsyncEffect(checkLoggedIn);
 
   return (
     <Main
